Convert adapters.
