+++ packages/adapter-jss/tests/JSSAdapter.test.ts
@@ -1,7 +1,8 @@
 /* eslint-disable jest/expect-expect */
 
-import { createRenderer } from 'fela';
-import webPreset from 'fela-preset-web';
+import { create } from 'jss';
+// @ts-ignore
+import preset from 'jss-preset-default';
 import { GLOBAL_STYLE_NAME } from 'aesthetic';
 import {
   cleanupStyleElements,
@@ -25,24 +26,28 @@ import {
   SYNTAX_PROPERTIES,
   SYNTAX_PSEUDO,
   SYNTAX_SUPPORTS,
-  FONT_ROBOTO,
-  FONT_CIRCULAR_MULTIPLE,
+  SYNTAX_VIEWPORT,
+  SYNTAX_CHARSET,
+  SYNTAX_IMPORT,
+  SYNTAX_IMPORT_MULTIPLE,
   SYNTAX_MEDIA_QUERY_NESTED,
   SYNTAX_KEYFRAMES_INLINE,
+  KEYFRAME_SLIDE_PERCENT,
   SYNTAX_FONT_FACES_INLINE,
   SYNTAX_RAW_CSS,
 } from 'aesthetic/lib/testUtils';
-import FelaAesthetic from '../src/FelaAesthetic';
+import JSSAdapter from '../src/JSSAdapter';
 
-describe('FelaAesthetic', () => {
-  let instance: FelaAesthetic<{}>;
+jest.mock('uuid/v4', () => () => 'uuid');
+
+describe('JSSAdapter', () => {
+  let instance: JSSAdapter;
 
   beforeEach(() => {
-    instance = new FelaAesthetic(
-      createRenderer({
-        plugins: [...webPreset],
-      }),
-    );
+    const jss = create();
+    jss.setup(preset());
+
+    instance = new JSSAdapter(jss);
   });
 
   afterEach(() => {
@@ -52,12 +57,49 @@ describe('FelaAesthetic', () => {
   DIRECTIONS.forEach(dir => {
     describe(`${dir.toUpperCase()}`, () => {
       it('converts and transforms inline styles', () => {
-        expect(instance.transformStyles([{ margin: 0 }, { padding: 2 }], { dir })).toBe('a b');
+        expect(instance.transformStyles(['foo', { margin: 0 }, { padding: 2 }], { dir })).toBe(
+          dir === 'ltr'
+            ? 'foo inline-0-0-1-1 inline-1-0-1-2'
+            : 'foo inline-0-0-25-1 inline-1-0-25-2',
+        );
+        expect(getFlushedStyles()).toMatchSnapshot();
+
+        // @ts-ignore Allow null
+        expect(instance.transformStyles(['foo', null], { dir })).toBe('foo');
+        expect(getFlushedStyles()).toMatchSnapshot();
       });
 
       describe('global sheet', () => {
         it('flushes and purges global styles from the DOM', () => {
-          renderAndExpect(instance, SYNTAX_GLOBAL, {}, { dir, global: true });
+          renderAndExpect(
+            instance,
+            SYNTAX_GLOBAL,
+            {
+              '@global': {
+                body: { margin: 0 },
+                html: { height: '100%' },
+                a: {
+                  color: 'red',
+                  '&:hover': {
+                    color: 'darkred',
+                  },
+                  '&:focus': {
+                    color: 'lightred',
+                  },
+                },
+                ul: {
+                  margin: 0,
+                  '&> li': {
+                    margin: 0,
+                  },
+                  '@media (max-width: 500px)': {
+                    margin: 20,
+                  },
+                },
+              },
+            },
+            { dir, global: true },
+          );
 
           instance.purgeStyles(GLOBAL_STYLE_NAME);
 
@@ -65,47 +107,95 @@ describe('FelaAesthetic', () => {
         });
 
         it('handles @font-face', () => {
-          const spy = jest.spyOn(instance.fela, 'renderFont');
-
-          instance.syntax.convertGlobalSheet(SYNTAX_FONT_FACE, { dir });
-
-          expect(spy).toHaveBeenCalledWith('Roboto', FONT_ROBOTO.srcPaths, FONT_ROBOTO_FLAT_SRC);
-          expect(spy).toHaveBeenCalledTimes(1);
+          renderAndExpect(
+            instance,
+            SYNTAX_FONT_FACE,
+            {
+              '@font-face': [FONT_ROBOTO_FLAT_SRC],
+            },
+            { dir, global: true },
+          );
         });
 
         it('handles mixed @font-face', () => {
-          const spy = jest.spyOn(instance.fela, 'renderFont');
-
-          instance.syntax.convertGlobalSheet(SYNTAX_FONT_FACE_MIXED, { dir });
-
-          expect(spy).toHaveBeenCalledWith('Roboto', FONT_ROBOTO.srcPaths, FONT_ROBOTO_FLAT_SRC);
-          expect(spy).toHaveBeenCalledWith(
-            'Circular',
-            FONT_CIRCULAR_MULTIPLE[0].srcPaths,
-            FONT_CIRCULAR_MULTIPLE_FLAT_SRC[0],
+          renderAndExpect(
+            instance,
+            SYNTAX_FONT_FACE_MIXED,
+            {
+              '@font-face': [FONT_ROBOTO_FLAT_SRC, ...FONT_CIRCULAR_MULTIPLE_FLAT_SRC],
+            },
+            { dir, global: true },
           );
-          expect(spy).toHaveBeenCalledTimes(5);
         });
 
         it('handles multiple @font-face', () => {
-          const spy = jest.spyOn(instance.fela, 'renderFont');
-
-          instance.syntax.convertGlobalSheet(SYNTAX_FONT_FACE_MULTIPLE, { dir });
-
-          expect(spy).toHaveBeenCalledWith(
-            'Circular',
-            FONT_CIRCULAR_MULTIPLE[0].srcPaths,
-            FONT_CIRCULAR_MULTIPLE_FLAT_SRC[0],
+          renderAndExpect(
+            instance,
+            SYNTAX_FONT_FACE_MULTIPLE,
+            {
+              '@font-face': FONT_CIRCULAR_MULTIPLE_FLAT_SRC,
+            },
+            { dir, global: true },
           );
-          expect(spy).toHaveBeenCalledTimes(4);
         });
 
         it('handles @keyframes', () => {
-          const spy = jest.spyOn(instance.fela, 'renderKeyframe');
+          renderAndExpect(
+            instance,
+            SYNTAX_KEYFRAMES,
+            {
+              '@keyframes fade': KEYFRAME_FADE,
+            },
+            { dir, global: true },
+          );
+        });
 
-          instance.syntax.convertGlobalSheet(SYNTAX_KEYFRAMES, { dir });
+        it('handles @viewport', () => {
+          renderAndExpect(
+            instance,
+            SYNTAX_VIEWPORT,
+            {
+              '@viewport': {
+                width: 'device-width',
+                orientation: 'landscape',
+              },
+            },
+            { dir, global: true },
+          );
+        });
 
-          expect(spy).toHaveBeenCalledWith(expect.anything(), {});
+        it('handles @charset', () => {
+          renderAndExpect(
+            instance,
+            SYNTAX_CHARSET,
+            {
+              '@charset': '"utf8"',
+            },
+            { dir, global: true },
+          );
+        });
+
+        it('handles single @import', () => {
+          renderAndExpect(
+            instance,
+            SYNTAX_IMPORT,
+            {
+              '@import': ['url("./some/path.css")'],
+            },
+            { dir, global: true },
+          );
+        });
+
+        it('handles multiple @import', () => {
+          renderAndExpect(
+            instance,
+            // @ts-ignore
+            SYNTAX_IMPORT_MULTIPLE,
+            {
+              '@import': ['url("./some/path.css")', 'url("./another/path.css")'],
+            },
+            { dir, global: true },
+          );
         });
       });
 
@@ -121,11 +211,6 @@ describe('FelaAesthetic', () => {
         });
 
         it('converts unified syntax to native syntax and transforms to a class name', () => {
-          instance.fela.renderFont('Roboto', FONT_ROBOTO.srcPaths, FONT_ROBOTO);
-
-          // @ts-ignore
-          instance.keyframes.fade = instance.fela.renderKeyframe(() => KEYFRAME_FADE, {});
-
           renderAndExpect(
             instance,
             SYNTAX_UNIFIED_LOCAL_FULL,
@@ -146,13 +231,13 @@ describe('FelaAesthetic', () => {
                 backgroundColor: '#337ab7',
                 verticalAlign: 'middle',
                 color: 'rgba(0, 0, 0, 0)',
-                animationName: 'k1',
+                animationName: 'fade',
                 animationDuration: '.3s',
-                ':hover': {
+                '&:hover': {
                   backgroundColor: '#286090',
                   borderColor: '#204d74',
                 },
-                '::before': {
+                '&::before': {
                   content: '"★"',
                   display: 'inline-block',
                   verticalAlign: 'middle',
@@ -172,7 +257,19 @@ describe('FelaAesthetic', () => {
         });
 
         it('handles attribute selectors', () => {
-          renderAndExpect(instance, SYNTAX_ATTRIBUTE, SYNTAX_ATTRIBUTE, { dir });
+          renderAndExpect(
+            instance,
+            SYNTAX_ATTRIBUTE,
+            {
+              attr: {
+                display: 'block',
+                '&[disabled]': {
+                  opacity: 0.5,
+                },
+              },
+            },
+            { dir },
+          );
         });
 
         it('handles descendant selectors', () => {
@@ -183,7 +280,7 @@ describe('FelaAesthetic', () => {
               list: {
                 margin: 0,
                 padding: 0,
-                '> li': {
+                '&> li': {
                   listStyle: 'bullet',
                 },
               },
@@ -193,7 +290,22 @@ describe('FelaAesthetic', () => {
         });
 
         it('handles pseudo selectors', () => {
-          renderAndExpect(instance, SYNTAX_PSEUDO, SYNTAX_PSEUDO, { dir });
+          renderAndExpect(
+            instance,
+            SYNTAX_PSEUDO,
+            {
+              pseudo: {
+                position: 'fixed',
+                '&:hover': {
+                  position: 'static',
+                },
+                '&::before': {
+                  position: 'absolute',
+                },
+              },
+            },
+            { dir },
+          );
         });
 
         it('handles multiple selectors (comma separated)', () => {
@@ -203,9 +315,9 @@ describe('FelaAesthetic', () => {
             {
               multi: {
                 cursor: 'pointer',
-                ':disabled': { cursor: 'default' },
-                '[disabled]': { cursor: 'default' },
-                '> span': { cursor: 'default' },
+                '&:disabled': { cursor: 'default' },
+                '&[disabled]': { cursor: 'default' },
+                '&> span': { cursor: 'default' },
               },
             },
             { dir },
@@ -217,11 +329,13 @@ describe('FelaAesthetic', () => {
             instance,
             SYNTAX_KEYFRAMES_INLINE,
             {
+              '@keyframes slide': KEYFRAME_SLIDE_PERCENT,
+              '@keyframes keyframe-1': KEYFRAME_FADE,
               single: {
-                animationName: 'k1',
+                animationName: 'slide',
               },
               multiple: {
-                animationName: 'k1, unknown, k2',
+                animationName: 'slide, unknown, keyframe-1',
               },
             },
             { dir },
@@ -233,6 +347,11 @@ describe('FelaAesthetic', () => {
             instance,
             SYNTAX_FONT_FACES_INLINE,
             {
+              '@font-face': [
+                FONT_ROBOTO_FLAT_SRC,
+                ...FONT_CIRCULAR_MULTIPLE_FLAT_SRC,
+                FONT_ROBOTO_FLAT_SRC,
+              ],
               single: {
                 fontFamily: 'Roboto',
               },
@@ -250,9 +369,14 @@ describe('FelaAesthetic', () => {
             SYNTAX_FALLBACKS,
             {
               fallback: {
-                background: ['red', 'linear-gradient(...)'],
-                display: ['block', 'inline-block', 'flex'],
-                color: ['blue'],
+                background: 'linear-gradient(...)',
+                display: 'flex',
+                fallbacks: [
+                  { background: 'red' },
+                  { display: 'block' },
+                  { display: 'inline-block' },
+                  { color: 'blue' },
+                ],
               },
             },
             { dir },
